Crate konnektoren_core

Modules

certificates

Module for creating certificates for the Konnektoren.

challenges

This module contains all the challenges that can be solved by the user.

commands

This module contains the command pattern implementation for the game.

game

Game module.

player_profile

Player profile module.

prelude

This is a prelude module that re-exports the most important types and traits.

session

Type Aliases

Xp

experience points