konnektoren_core::challenges::solvable

Trait Solvable

pub trait Solvable {
    // Required method
    fn solve(&mut self, input: ChallengeInput) -> Result<bool>;
}

Required Methods

fn solve(&mut self, input: ChallengeInput) -> Result<bool>

Attempts to solve a part of the challenge with the given input. Returns true if the part is successfully solved, otherwise false.